+++ fission-cli/library/Fission/CLI/IPFS/Configure.hs
@@ -97,7 +97,7 @@ setSwarmAddresses = do
     , "Addresses.Swarm"
     ]
     (fromString . show $ encode ("/ip4/0.0.0.0/tcp/4231/ws" : existing))
-               -- ^^^^ `show` to get string escaping
+               -- `show` to get string escaping
 
 enableRelay ::
   ( MonadLocalIPFS m
