+++ fission-core/library/Fission/URL/Validation.hs
@@ -11,60 +11,7 @@ import           Fission.Prelude
 
 import qualified Fission.Security as Security
 
--- $setup
--- >>> :set -XOverloadedStrings
-
 -- | Confirm that a raw is valid
---
--- >>> isValid "simple"
--- True
---
--- >>> isValid "happy-name"
--- True
---
--- >>> isValid "under_score"
--- True
---
--- Blocklisted words are not allowed
---
--- >>> isValid "recovery"
--- False
---
--- They're not case sensitive
---
--- >>> isValid "reCovErY"
--- False
---
--- They can't contain uppercase characters at all
---
--- >>> isValid "hElLoWoRlD"
--- False
---
--- Nor are various characters
---
--- >>> isValid "plus+plus"
--- False
---
--- >>> isValid "-startswith"
--- False
---
--- >>> isValid "endswith-"
--- False
---
--- >>> isValid "_startswith"
--- False
---
--- >>> isValid "with.space"
--- False
---
--- >>> isValid "with.dot"
--- False
---
--- >>> isValid "has.two.dots"
--- False
---
--- >>> isValid "name&with#chars"
--- False
 isValid :: Text -> Bool
 isValid txt =
   all (== True) preds
