+++ fission-web-server/library/Fission/Web/Server/User/Role/Types.hs
@@ -1,6 +1,7 @@
 module Fission.Web.Server.User.Role.Types (Role (..)) where
 
-import           Database.Persist.TH
+import           Database.Persist.Sql
+import qualified RIO.Text             as Text
 
 import           Fission.Prelude
 
@@ -12,4 +13,16 @@ data Role
            , Eq
            )
 
-derivePersistField "Role"
+instance PersistField Role where
+  toPersistValue role =
+    PersistText . Text.pack $ show role
+
+  fromPersistValue val =
+    case fromPersistValueText val of
+      Right "Regular" -> Right Regular
+      Right "Admin"   -> Right Admin
+      Right _         -> Left "Not a valid Role value"
+      Left err        -> Left err
+
+instance PersistFieldSql Role where
+  sqlType _ = SqlString
