+++ hie.yaml
@@ -1,0 +1,58 @@
+cradle:
+  stack:
+    - path: "fission-cli/library"
+      component: "fission-cli:lib"
+
+    - path: "fission-cli/cli/Main.hs"
+      component: "fission-cli:exe:fission"
+
+    - path: "fission-cli/cli/Paths_fission_cli.hs"
+      component: "fission-cli:exe:fission"
+
+    - path: "fission-cli/benchmark/Main.hs"
+      component: "fission-cli:bench:fission-cli-benchmark"
+
+    - path: "fission-cli/benchmark/Paths_fission_cli.hs"
+      component: "fission-cli:bench:fission-cli-benchmark"
+
+    - path: "fission-core/library"
+      component: "fission-core:lib"
+
+    - path: "fission-core/library"
+      component: "fission-core:test:fission-core-test"
+
+    - path: "fission-core/test"
+      component: "fission-core:test:fission-core-test"
+
+    - path: "fission-core/benchmark/Main.hs"
+      component: "fission-core:bench:fission-core-benchmark"
+
+    - path: "fission-core/benchmark/Paths_fission_core.hs"
+      component: "fission-core:bench:fission-core-benchmark"
+
+    - path: "fission-web-api/library"
+      component: "fission-web-api:lib"
+
+    - path: "fission-web-client/library"
+      component: "fission-web-client:lib"
+
+    - path: "fission-web-server/library"
+      component: "fission-web-server:lib"
+
+    - path: "fission-web-server/server/Main.hs"
+      component: "fission-web-server:exe:fission-server"
+
+    - path: "fission-web-server/server/Paths_fission_web_server.hs"
+      component: "fission-web-server:exe:fission-server"
+
+    - path: "fission-web-server/library"
+      component: "fission-web-server:test:fission-web-server-test"
+
+    - path: "fission-web-server/test"
+      component: "fission-web-server:test:fission-web-server-test"
+
+    - path: "fission-web-server/benchmark/Main.hs"
+      component: "fission-web-server:bench:fission-web-server-benchmark"
+
+    - path: "fission-web-server/benchmark/Paths_fission_web_server.hs"
+      component: "fission-web-server:bench:fission-web-server-benchmark"
